==> 1-Load data.Rmd <==
---
title: "Transplantation study: load data"
author: "Giulia"
date: "1/26/2022"
output: html_document
---

#Load packages
```{r, load packages}
#R version in use: version.string R version 3.6.1 (2019-07-05)

#load pachages
library(ampvis2) #version 2.5.1
library(tidyverse) #version 1.2.1
```

#Load data
```{r Load raw data into R}
#import ASV table and MiDAS 3 taxonomy: "Additional file 2"
ASVtableSintax <- data.table::fread("C:/Users/Giulia/Documents/R/Immigration/Transplantation/paper_codes/data_raw/Additional-file-2_ASVtableSintax.tsv", header = T) #137 columns (as samples) + 8 taxonomic level including ASV level

#import sample metadata: "Additional file 3"
mymetadata <- readxl::read_excel("C:/Users/Giulia/Documents/R/Immigration/Transplantation/paper_codes/data_raw/Additional-file-3_metadata.xlsx.")

#Create an ampvis object that merges ASV table containing MiDAS 3 taxonomy and sample metadata
d_amp <- amp_load(otutable = ASVtableSintax, 
              metadata = mymetadata)
```

#Subset

```{r Subset samples}
#remove control samples and samples with total read counts lower than 10.000 

d_amp_sub <- amp_subset_samples(d_amp, 
                              minreads = 10000, 
                              normalise = F, #OBS: here we Do NOT normalise the read counts 
                              !SampleID %in% c("MQ171221-130", "MQ171221-131"))  #remove positive and negative control samples

```

#Long format 

The dataset is converted in long format. The read counts are NOT normalized to sample yet.
```{r}
#subset and long format
d_long_reads <-  amp_export_long(d_amp_sub,
                            metadata_vars = c("SampleID", "Plants", "SampleType", "Date", "Days", "Transplantation")) #select needed columns

#add hybrid taxonomy column for unclassified species. If species level classification is present, the species name is pasted into a new column "Species_rename", otherwise the ASV is combined with the first available level of classification (e.g. Genus, or Family, or Order, etc.)
d_long_reads <- d_long_reads %>%
  mutate(Species_rename = ifelse(Kingdom =="", paste0("k__",OTU),
                            ifelse(Phylum == "", paste0(Kingdom,"__",OTU),
                                 ifelse(Class =="", paste0(Phylum,"__",OTU),
                                        ifelse(Order =="",paste0(Class,"__",OTU),
                                               ifelse(Family =="",paste0(Order,"__",OTU),
                                                      ifelse(Genus == "",paste0(Family,"__", OTU),
                                                             ifelse(Species =="", paste0(Genus,"__",OTU),
                                                                    Species))))))))
```

###aggregate read counts
```{r}
#aggregate read counts at species_rename level 
library(dplyr)

d_long_reads2 <- d_long_reads %>%
                      group_by(SampleID) %>%
                      mutate(SampleReads = sum(count)) %>% #to create a column with total reads within sample
                      ungroup() %>%
                      group_by(SampleID, Species_rename) %>% #group classified and unclassified species by Sample
                      mutate(count_sp = sum(count)) %>% 
                      dplyr::select(-OTU, -count, -Species) %>% #remove columns that refer only to classified species
  ungroup() %>%                     
  unique()
```

####check read counts

Check that the aggregation in the new dataset retuns correct number of read counts at species level from the previous dataset
```{r}
#read counts of one specific species from one specific sample from the dataset with non aggregated counts
check1 <- d_long_reads %>%
  group_by(Plants, SampleType, Date, Species) %>%
  dplyr::mutate(count_sp = sum(count)) %>%
  subset(., Species %in% "s__midas_s_5" & Plants %in% "Donor" & Days %in% "-140" & SampleType %in% "Activated Sludge")


#read counts in one species in one sample from datatet with aggregated counts 
check2 <- subset(d_long_reads2, Species_rename %in% "s__midas_s_5" & Plants %in% "Donor" & Days %in% "-140" & SampleType %in% "Activated Sludge")

#check differences
setdiff(sum(check1$count), check2$count_sp)
```
the difference should be zero.

###calculate relative read abundance
```{r}
d_long <- d_long_reads2 %>%
  group_by(SampleID, Species_rename) %>% 
  mutate(rra_sp = (count_sp/SampleReads)*100) %>% 
  ungroup()
```


#Rename SampleType
```{r}
d_long2 <- d_long #create a backup dataset

#rename SampleType names with abbreviations
d_long2$SampleType[d_long2$SampleType %in% "Wastewater influent"] <- "IWW"
d_long2$SampleType[d_long2$SampleType %in% "Activated Sludge"] <- "AS"
```

#Create Sample Group
```{r}
#add sample group column based on Plants and time of transplantation. Labels as in Table S2
d_long2$SampleGroup[d_long2$SampleType %in% "IWW" & d_long2$Plants %in% "Donor"] <- "D-IWW"
d_long2$SampleGroup[d_long2$SampleType %in% "IWW" & d_long2$Plants %in% "Recipient"] <- "R-IWW"

d_long2$SampleGroup[d_long2$SampleType %in% "AS" & d_long2$Plants %in% "Donor" & d_long2$Days < 0] <- "D-AS(before)"
d_long2$SampleGroup[d_long2$SampleType %in% "AS" & d_long2$Plants %in% "Recipient" & d_long2$Days < 0] <- "R-AS(before)"

d_long2$SampleGroup[d_long2$SampleType %in% "AS" & d_long2$Plants %in% "Donor" & d_long2$Days <= 40 & d_long2$Days >= 0] <- "D-AS(after-early)"
d_long2$SampleGroup[d_long2$SampleType %in% "AS" & d_long2$Plants %in% "Recipient" & d_long2$Days <= 40 & d_long2$Days >= 0] <- "R-AS(after-early)"

d_long2$SampleGroup[d_long2$SampleType %in% "AS" & d_long2$Plants %in% "Donor" & d_long2$Days <= 150 & d_long2$Days > 40] <- "D-AS(after-intermediate)"
d_long2$SampleGroup[d_long2$SampleType %in% "AS" & d_long2$Plants %in% "Recipient" & d_long2$Days <= 150 & d_long2$Days > 40] <- "R-AS(after-intermediate)"

d_long2$SampleGroup[d_long2$SampleType %in% "AS" & d_long2$Plants %in% "Donor" & d_long2$Days > 150] <- "D-AS(after-late)"
d_long2$SampleGroup[d_long2$SampleType %in% "AS" & d_long2$Plants %in% "Recipient" & d_long2$Days > 150] <- "R-AS(after-late)"


#Check nr of samples within each sample group, as in Table S2
d_long2 %>% 
  ungroup() %>% 
  select(-Species_rename) %>% 
  unique() %>% 
  distinct(SampleID, .keep_all = T) %>% 
  group_by(SampleGroup) %>% 
  count()
```
#Dataset Stats

```{r}
#nr of samples per group
d_long2 %>% 
  ungroup() %>% 
  group_by(Plants, SampleType) %>% 
  select(Days) %>% 
  unique() %>% 
  count()

#nr of species per sample
d_long2 %>% 
   ungroup() %>% 
  group_by(SampleID, Plants, SampleType, Days) %>% 
  select(Species_rename) %>% 
  unique() %>% 
  count()

#Total reads per sample
sr_stats <- d_long2 %>% 
  ungroup() %>% 
  select(SampleID, Plants, SampleType, Days, SampleReads) %>% 
  unique()

summary(sr_stats$SampleReads)
sd(sr_stats$SampleReads)
```

#Adjust Species names

Remove underscores from taxonomy name
```{r}
d_long3 <- d_long2 #backup of dataset

#paste Genus classification for classified species
d_long3$Genus2 <- ifelse(!grepl("ASV", d_long3$Species_rename, fixed = F),
                            paste(d_long3$Genus),
                            NA)
## remove underscore from Genus classification
d_long3$Genus2 <- ifelse(grepl("g__", d_long3$Genus2, fixed = F),
                      gsub("g__", "", d_long3$Genus2),
                     NA)

#create a temporary species name where taxonomic info from unclassified species are separated and removed
d_long3 <- d_long3 %>% 
  mutate(Species_renameB = paste0(Species_rename)) %>% 
  separate(., Species_renameB, c("Root", "Taxa", "ASV"), sep = "__") %>% #split column
  dplyr::select(-Root) #remove unneeded column

#for unclassified species, paste the taxonomic level available without any underscores
d_long3$Species_rename2 <- ifelse(is.na(d_long3$Genus2),
                               paste(d_long3$Taxa, d_long3$ASV, sep = " "),
                               paste(d_long3$Genus2, d_long3$Taxa, sep = " "))

#remove columns that won't be used
d_long4 <- d_long3 %>% 
  select(-c(Kingdom:Family), -ASV)
```


#Export 

```{r, eval = FALSE}
#Export ampvis-format dataset for further analysis
##normalise read counts to relative read abundance within sample before exporting
d_amp_sub2 <- amp_subset_samples(d_amp_sub, normalise = T)
saveRDS(object = d_amp_sub2, file = "C:/Users/Giulia/Documents/R/Immigration/Transplantation/paper_codes/data_generated/d_amp_sub.rds")

#Export long-format dataset for further analysis
data.table::fwrite(d_long4,file = "C:/Users/Giulia/Documents/R/Immigration/Transplantation/data_generated/d_long4.tsv", sep = "\t")
```
